requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
requests==2.31.0
audioop-lts==0.2.1
